refactor: Room에 socket event와 관련된 것은 RoomEvent로 분리
