+++ src/rooms/item/index.ts
@@ -1,52 +1,25 @@
-import { SocketInformation } from 'src/conversation-events/interfaces/socket-information.interface';
 import { ChatData } from 'src/conversation-datas/data/chatting';
-import { RoomSocket } from 'src/conversation-events/interfaces/room-socket.interface';
 import { User } from 'src/users/entities/user.entity';
-import { Logger } from '@nestjs/common';
-
-export const logger = new Logger('Room');
-
-export enum RoomStatus {
-  WATING = 'Wating',
-  INVITING = 'Inviting',
-  RUNNING = 'Running',
-  CREATOR_OUT = 'CreatorOut',
-  PARTICIPANT_OUT = 'ParticipantOut',
-  CLOSING = 'Closing',
-}
+import { RoomEvent } from './room-event';
 
 export class Room {
-  uuid: string;
-  status: RoomStatus;
   creatorPk: number;
   participantPk: number;
-  creatorSocket: RoomSocket;
-  participantSocket: RoomSocket;
-  watingSockets: RoomSocket[];
   startedAt: Date;
   finishedAt: Date;
   prUrl: string;
-  globalTimeoutId: NodeJS.Timeout;
-  outSocketInformation: SocketInformation;
+
+  roomEvent: RoomEvent;
   data: RoomData;
 
   constructor(uuid: string, creator: User, prUrl: string) {
-    this.uuid = uuid;
-    this.status = RoomStatus.WATING;
     this.creatorPk = creator.pk;
     this.prUrl = prUrl;
     this.startedAt = new Date();
-    this.watingSockets = [];
+    this.roomEvent = new RoomEvent(uuid);
     this.data = {
       chat: new ChatData(),
     };
-  }
-
-  clearTimeout() {
-    logger.log(`${this.uuid}의 timeout이 제거되었습니다.`);
-    clearTimeout(this.globalTimeoutId);
-    this.globalTimeoutId = undefined;
-    this.outSocketInformation?.clearTimeout();
   }
 }
